Deduplicate fixtures
