+++ tests/test_custom_extensions_in_hooks.py
@@ -21,14 +21,6 @@ def template(request):
     return 'tests/test-extensions/' + request.param
 
 
-@pytest.fixture
-def output_dir(tmp_path):
-    """Fixture. Create and return custom temp directory for test."""
-    output_path = tmp_path.joinpath("hello")
-    output_path.mkdir()
-    return str(output_path)
-
-
 @pytest.fixture(autouse=True)
 def modify_syspath(monkeypatch):
     """Fixture. Make sure that the custom extension can be loaded."""
